requirements.txt
# Webook Pro Bot Dependencies
# Enterprise-grade Telegram ticket booking system

# Telegram
aiogram>=3.3.0

# Browser Automation
playwright>=1.40.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Async File I/O
aiofiles>=23.0.0

# HTML Parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# YAML Config
pyyaml>=6.0.0
